Fix table animation: in progress
